fix: use first UTF-8 byte to determine length of a character
The UTF-8 decoding logic was using the last byte read to determine the length of
the current character being decoded. This meant that the only characters it
could successfully decode were 1-byte and 4-byte UTF-8. Characters of other
lengths that have 0b10-prefixed continuation bytes would be incorrectly
interpreted as meaning the character was 4 bytes long.