fix: treat all SGR resets for unsupported modes as no-ops
The CI environments send `<esc>[27m` where local execution does not. We do not
support the mode this is resetting, so just treat this and other similar
sequences as things we can safely ignore.

There is still an issue affecting the 4-byte UTF-8 test case on Linux that I
have not yet managed to debug.