fix: clamp terminal dimensions, not file dimensions
There are three relevant rows×columns dimensions when parsing a file:

  1. The dimensions of the file itself;
  2. The dimensions of the virtual terminal being painted; and
  3. The dimensions of the sliding window into the file being read.

This code was incorrectly adjusting (1) up and down and then giving some
combination of these and (3) to `run_vim` as if these were the terminal
dimensions. A result of this was that `run_vim` would render a different,
smaller window onto the virtual terminal than that expected by `vimcat_read`.
This, in combination with the bug fixed in the prior commit, resulted in lines
being skipped when dealing with a >1000 line file.