remove stub of support for combining marks
The original plan was to support an arbitrary number of non-spacing combining
marks by using the alternate heap-allocated buffer for a grapheme. However on
investigating this, I discovered there are ~2K non-spacing combining marks,
spread across multiple Unicode blocks, with sources disagreeing on which
characters belong in this category. The ground truth is the non-spacing mark
(`Mn`) category in the Unicode specification, but handling this correctly still
remains much more complex than I anticipated.

The cost of the implementation in terms of both complexity and runtime
performance overhead does not seem worth it. Getting this right would be
challenging and using a more principled UTF-8 approach like ICU would be a
significant change of design. Since the primary use case of `vimcat` is
highlighting program source code, non-spacing combining marks are unlikely to
appear. Furthermore the only infidelity is in the virtual terminal’s notion of
current position. That is, to observe a problem here your input source to the
terminal needs to either (1) have lines spanning more than Vim’s 10000 column
limit or (2) have ANSI sequences that jump backwards into lines containing such
marks. I have not observed Vim emit anything like (2).