fix: retain “bright”/aixterm colour directive
When seeing a Control Sequence Identifier (CSI) identifying a “bright” colour
(e.g. `<esc>[91m`, enabling bright red foreground), the virtual terminal would
recognise the sequence but fail to note a custom colour was being set. The
effect of this would be to enable bold, but leave the colour set to the default
if it was so previously.