support “bright”/aixterm colours as 256-bit colours
Instead of mapping the “bright” colours to a bold version of their equivalent
8-bit colour, they are now mapped into 256-bit colour space, preserving their
original properties.

A side effect of this is that an input sequence like `<esc>[91m` no longer
emerges as an output sequence like `<esc>[31;1m` but rather as a 256-bit
sequence like `<esc>[38;5;9m`. A future change change will improve on both of
these, by supporting round tripping this to the original `<esc>[91m`.