namespace 'debug'
When building libvimcat as a shared library, the
`__attribute__((visibility("internal")))` annotation correctly made `debug`
invisible to external users. But when building as a static library, it has no
effect. The result of this is that users can get link errors if they have their
own `debug` symbol.

There are various workarounds to this (“pre-linking” on macOS,
`objcopy --localize-hidden`, …), but it seems simpler to just rename this
symbol to something more likely to be unique.