read_core: optimise file extent calculation by stopping early
When we have been requested to highlight a particular line in a file, we do not
actually care about the precise number of lines in the file. We just care that
the number of lines is ≥ the line number requested. So this change makes the
extent calculation bail out early if we have already figured that out.