fix pipe file descriptors polluting other threads
This code creates a pipe, but was not unconditionally setting close-on-exec. The
problem with this is that another thread may be currently running `fork`+`exec`.
The newly started process by this `exec` (incorrectly) acquires copies of these
file descriptors. We cannot control concurrency here because this is library
code that may be called from a multithreaded context. In fact Clink¹ does
exactly this.

The intuitive fix would be to `fcntl` both ends of the pipe to set
close-on-exec. But this is racy, as explained in a recent Zig commit.² There
seems no way to avoid this race on macOS, but at least we can address it on
Linux and the BSDs.

¹ https://github.com/smattr/clink
² https://github.com/ziglang/zig/commit/389181f6be8810b5cd432e236a962229257a5b59