set close-on-exec on /dev/null file descriptor
Similar to the issue discussed in the prior commit, another thread can be
attempting `exec` while this is running and its child will incorrectly inherit
an open file descriptor to /dev/null. The problem is less severe in this case,
because there are not many problems a child can cause with an unintended
/dev/null file descriptor. But we still should not be polluting their descriptor
space.

Glibc’s `open` man page has a good description of the general problem:

  Note that the use of this flag is essential in some multithreaded programs,
  because using a separate `fcntl`(2) `F_SETFD` operation to set the
  `FD_CLOEXEC` flag does not suffice to avoid race conditions where one thread
  opens a file descriptor and attempts to set its close-on-exec flag using
  `fcntl`(2) at the same time as another thread does a `fork`(2) plus
  `execve`(2). Depending on the order of execution, the race may lead to the
  file descriptor returned by `open`() being unintentionally leaked to the
  program executed by the child process created by `fork`(2). (This kind of race
  is in principle possible for any system call that creates a file descriptor
  whose close-on-exec flag should be set, and various other Linux system calls
  provide an equivalent of the `O_CLOEXEC` flag to deal with this problem.)

As with the previous commit, it seems macOS has no solution to this problem.